--- deploy/control/actions.py
@@ -1,4 +1,0 @@
-from deploy import exceptions
-
-from deploy.conf import StackConf
-
