+++ deploy_raw/conf/core.py
@@ -1,20 +1,20 @@
 from __future__ import annotations
 
-from deploy import exceptions
+from deploy_raw import exceptions
 import re
 import yaml
 from pathlib import Path
 
-from deploy.conf.service import (
+from deploy_raw.conf.service import (
 	ServiceConf,
 	S
 )
-from deploy.conf.service import (
+from deploy_raw.conf.service import (
 	SiteConf,
 	PostgresConf,
 )
 
-from deploy.control.mixins import DockerBuildMixin
+from deploy_raw.control.mixins import DockerBuildMixin
