+++ deploy_raw/conf/service/parts.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from deploy import exceptions
+from deploy_raw import exceptions
 from pathlib import Path, PosixPath
