+++ deploy_raw/control/convert.py
@@ -2,11 +2,11 @@
 Not sure if this will be used
 """
 
-from deploy import exceptions
+from deploy_raw import exceptions
 
-from deploy.conf import StackConf
-from deploy.conf.service import ServiceConf
-from deploy.conf.service.parts import (
+from deploy_raw.conf import StackConf
+from deploy_raw.conf.service import ServiceConf
+from deploy_raw.conf.service.parts import (
 	Volume
 )
