+++ deploy_raw/control/core.py
@@ -1,8 +1,8 @@
 from __future__ import annotations
 
-from deploy import exceptions
+from deploy_raw import exceptions
 
-from deploy.conf import StackConf
+from deploy_raw.conf import StackConf
